Refactor naming and add comments
This commit only tangentially alters functionality. It mainly changes
the names of methods and variables while adding substantial comments.
It uses Black as formatter.

The comments focus the response format of GTO+. These responses are
dense and require non-trivial parsing. The comments should make it
easier to understand _why_ the code is written the way it is and, more
importantly, make it easier to update if GTO+ changes its format.

Other changes include moving or defining magic numbers and refactoring
methods according to their level of abstraction. For example, in
`request_node_data` there are now multiple `get` methods. These perform
the calls to GTO+ and parsing of the reponse. Because this code is at a
lower level of abstraction, it gets its own method. The intent is to
allow the reader to dive deeper only when they feel the need while
relying on naming to indicate _what_ something does rather than _how_.
If the reader is interested in the how, they can drop down a level of
abstraction to the helper method.

Lastly, this commit rearranges methods and uses underscores to indicate
intended visibility. That is, methods with an underscore are intended to
be private; they are helper methods. Publicly available methods
generally precede these helper methods.

In follow-up commits, I plan to add comments to methods currently marked
with TODOs, add type-hinting, and add logging levels to make future
development easier.